# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Hyperledger Interface Server",
    "title": "his",
    "version": "1.0.0"
  },
  "paths": {
    "/his/v0/api/app/{appid}/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "operationId": "getConsents",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/consent"
              }
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "tags": [
          "consents"
        ],
        "operationId": "createConsent",
        "parameters": [
          {
            "x-go-name": "Consent",
            "description": "The consent to submit.",
            "name": "consent",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/consent"
            }
          },
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "$ref": "#/definitions/consent"
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "tags": [
          "consents"
        ],
        "operationId": "deleteConsents",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/nocontent"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/his/v0/api/app/{appid}/consents/{consentid}": {
      "get": {
        "tags": [
          "consents"
        ],
        "operationId": "getConsent",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ConsentID",
            "description": "The consent ID to submit.",
            "name": "consentid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "$ref": "#/definitions/consent"
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "tags": [
          "consents"
        ],
        "operationId": "deleteConsent",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ConsentID",
            "description": "The consent ID to submit.",
            "name": "consentid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/nocontent"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/his/v0/api/app/{appid}/consumer/{consumerid}/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "operationId": "getConsents4Consumer",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ConsumerID",
            "description": "The consumer ID to submit.",
            "name": "consumerid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/consent"
              }
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/his/v0/api/app/{appid}/consumer/{consumerid}/owner/{ownerid}/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "operationId": "getConsents4ConsumerOwner",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ConsumerID",
            "description": "The consumer ID to submit.",
            "name": "consumerid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "OwnerIP",
            "description": "The owner ID to submit.",
            "name": "ownerid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/consent"
              }
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/his/v0/api/app/{appid}/isconsent": {
      "post": {
        "tags": [
          "consents"
        ],
        "operationId": "isConsent",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/isConsentResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/his/v0/api/app/{appid}/owner/{ownerid}/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "operationId": "getConsents4Owner",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AppID",
            "description": "The application ID to submit.",
            "name": "appid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "OwnerIP",
            "description": "The owner ID to submit.",
            "name": "ownerid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "consent",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/consent"
              }
            }
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/ocms/v3/api/version": {
      "get": {
        "tags": [
          "version"
        ],
        "operationId": "getVersion",
        "responses": {
          "200": {
            "$ref": "#/responses/versionResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    }
  },
  "definitions": {
    "consent": {
      "type": "object",
      "required": [
        "appid",
        "consentid",
        "ownerid",
        "consumerid"
      ],
      "properties": {
        "appid": {
          "type": "string",
          "x-go-name": "AppID"
        },
        "consentid": {
          "type": "string",
          "x-go-name": "ConsentID"
        },
        "consumerid": {
          "type": "string",
          "x-go-name": "ConsumerID"
        },
        "dataaccess": {
          "type": "string",
          "x-go-name": "DataAccess"
        },
        "datatype": {
          "type": "string",
          "x-go-name": "DataType"
        },
        "dtbegin": {
          "type": "string",
          "x-go-name": "Dt_begin"
        },
        "dtend": {
          "type": "string",
          "x-go-name": "Dt_end"
        },
        "ownerid": {
          "type": "string",
          "x-go-name": "OwnerID"
        }
      },
      "x-go-name": "Consent",
      "x-go-package": "github.com/pascallimeux/his/modules/ocms"
    }
  },
  "responses": {
    "genericError": {
      "description": "Genric error for API",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "in: path"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "isConsentResponse": {
      "description": "is the consent exist",
      "headers": {
        "isconsent": {
          "type": "string"
        }
      }
    },
    "nocontent": {
      "description": "no content",
      "headers": {
        "nocontent": {
          "type": "string"
        }
      }
    },
    "versionResponse": {
      "description": "Version of smart contract consent",
      "headers": {
        "version": {
          "type": "string"
        }
      }
    }
  }
}